requirements.txt
asgiref==3.4.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
